Strange shit
When loading an XNU Kernel Mach-O, the vmaddr props of segment and section structs
begin with 0xffffff, so for example the __TEXT segment->vmaddr is 0xfffffff007004000.
With Kernel files, all of the data constructed and printed matches that of otool and
MachOViewer.

However, there is strange shit going on with non-kernel mach-o's. Now this maybe due
to a lack of understanding of the file format on my part, but fuck knows what's going
on here.

First of all, addresses do not have the 0xff.... padding. Instead, the __TEXT segment
starts at 0x10000. This results in the sections being read completely wrong, because
the address is wrong.

There is a way, by adding 32 to the file offset prop, that you can read the sections,
but that breaks the kernel images. So, to read the section data using the vmaddr prop
one must do the following with the vmaddr:

    if (segment->vmaddr < 0xffffff0000000000) {
        uint64_t vmaddr = 0xffffff0000000010 + segment->vmaddr;
        segment->vmaddr = vmaddr;
    }

This *mostly* fixes things. I did notice, however, that properties such as the file
offset and align were off by a few bytes. I cannot imagine why this is, because the
rest of the structure appears to be in good shape.

Anyone more experienced with the Mach-O format, it would be a massive help if you could
give your thoughts.

Contact: Twitter @h3adsh0tzz