Multiple changes...
Firstly, I'm working on symbol and string table stuff, currently trying to
grab the correct string in the table for the current symbol. This is provin
g more difficult than i thought.

Nect, I'm trying to detect the Universal Binary format, with the magic 0xcafebabe,
but there are some issues with that

added another test file