The single header is now completely seperated...
Each area of the Mach-O format has its own file and header.

I have bundled segments and sections together, since they are closely related and the compiler
kept crying when i had them seperate.

Also, i'll change the names of the C files and remove the 'macho-' prefix, that was only while
i had the old files around.

The last thing to do is symbols, then im back to where i was when everything was in one header.