Sepsplit now works:
It appears that despite the Mach-O's inside the sep firmware image being 64bit, Apple
chose to use 32bit mach_header's. This means that because i was using 64bit headers,
the extra reserved uint32_t at the end of the 64bit header that the 32bit one doesn't
meant that the load command pointer in calc_size was off by sizeof(uint32_t)

now i use the 32bit header on the 64bit macho, it works fine. This required some changes
to libhelper-macho, but it resulted in compatibility with loader.h