Added support for arm64 (DTK)
Added support for libhelper to know whether it's being compiled
on an arm64 Mac, aka Developer Transition Kit (DTK). This does
not change the functionality of the library, although some changes
may need to be made for checking platform type, libhelper version
and architecture.

Currently in a development state, so not ready for release yet.
This version is currently being used in the arm64-dtk versions
of HTool and Img4helper.