libhelper-core: Add initial libhelper-core code.
This commit adds the initial libhelper-core functionality. The library
is split into "modules" which are intended to be built optionally, so
if the user does not require a Mach-O parser, they don't have to compile
it for their project - this is intended to be reflecting in the version
number also.

This commit contains the following libhelper-core functionality:
  - File handling API.
  - HLibc API (Strings and Singly-linked lists).
  - Boyermoore Horspool memmem().
  - Libhelper version API.

Also contained in this commit is a libhelper-core-test.

Libhelper can be compiled by:
$ mkdir build && cd build && cmake .. && make

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>