libhelper-macho: Implements some 64-bit parsing.
This commit adds some 64-bit parsing support, and the test application.

Basic load command parsing is also added. The next step is to add parsing
for induvidual load commands like Segment/section, library etc. This is
required to finish the 64-bit parser.

The 32-bit parser will be added, but is not the main priority as it's not
likely to be used much.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>