segment parsing: added definitions & initial parsing code
There seems to be an issue where if mach_section_64_load is called, it breaks the
data in macho->header, but not the data after it - as the data is continued to be
parsed correctly. The header is broken after parsing the first segment. This needs
to be fixed.

The issue is occuring despite the same code as v2.0.0 being used. Is this a compiler
issue?