FAT: Merge current FAT parser changes into libhelper-macho dev branch. (#9)
* FAT: Initial FAT/Universal binary parsing API.

The aim is for libhelper to have a more improved FAT/Universal binary
parser and API than before, one that can easily switch between archs.

Currently HTool has some experimental fat parsing, so the code from
that will be used to implement it in libhelper.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* Header: Added ARM CPU Extension handling.

The function `mach_header_get_cpu_descriptor()` has been added. When called,
it will return a string describing the CPU Architecture version and any
extensions. For example, a CPU which Apple describe as arm64e with Pointer
Authentication and Memory Tagging Extension would return the following:

	"ARMv8.5-A, MTE, PAC"

Whereas a simple arm64 CPU with no extensions, or a binary built with the
CPU type CPU_SUBTYPE_ARM64_ALL, the following would be returned:

	"ARMv8-A".

This functionality allows users to see which CPU architecture a binary was
compiled for. This is particularly useful for looking at Kernels.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* fat: Added fat parser test

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* header: fixed issue where arm64e was being returned as arm64_unknown

Co-authored-by: h3 <h3@h3s-iMac-Pro.local>