Implement Mach-O Support into libhelper v3.0.0 (#10)
* libhelper-macho: Implements some 64-bit parsing.

This commit adds some 64-bit parsing support, and the test application.

Basic load command parsing is also added. The next step is to add parsing
for induvidual load commands like Segment/section, library etc. This is
required to finish the 64-bit parser.

The 32-bit parser will be added, but is not the main priority as it's not
likely to be used much.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* libhelper-macho: Implements LC helpers & defines FAT archive

* libhelper-macho: type definitions for handling segment commands

* segment parsing: added definitions & initial parsing code

There seems to be an issue where if mach_section_64_load is called, it breaks the
data in macho->header, but not the data after it - as the data is continued to be
parsed correctly. The header is broken after parsing the first segment. This needs
to be fixed.

The issue is occuring despite the same code as v2.0.0 being used. Is this a compiler
issue?

* header: remove free(hdr)

This call to free hdr was causing the issue mentioned in the last commit. All
should be well now.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* segment: add architecture-agnostic parsing

* segment/section: add (almost) all parser code.

This should complete the segment and section parsers. Ideally the next things
to do is write out proper documentation in the headers and write proper tests
for all the commands.

* docs: improved documentation for libhelper-macho.h

* tests: implemented proper tests for libhelper-macho

* misc: update copyright in license header

* tools: added python script to check which load commands aren't implemented

* load-commands: implemented dylib commands and others

* tools: updated implementation checker, shows both implemented and not implemented structs

* load-commands: added sub_ commands, and dylinker

* cmake: fix directories for including as submodule

* FAT: Merge current FAT parser changes into libhelper-macho dev branch. (#9)

* FAT: Initial FAT/Universal binary parsing API.

The aim is for libhelper to have a more improved FAT/Universal binary
parser and API than before, one that can easily switch between archs.

Currently HTool has some experimental fat parsing, so the code from
that will be used to implement it in libhelper.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* Header: Added ARM CPU Extension handling.

The function `mach_header_get_cpu_descriptor()` has been added. When called,
it will return a string describing the CPU Architecture version and any
extensions. For example, a CPU which Apple describe as arm64e with Pointer
Authentication and Memory Tagging Extension would return the following:

	"ARMv8.5-A, MTE, PAC"

Whereas a simple arm64 CPU with no extensions, or a binary built with the
CPU type CPU_SUBTYPE_ARM64_ALL, the following would be returned:

	"ARMv8-A".

This functionality allows users to see which CPU architecture a binary was
compiled for. This is particularly useful for looking at Kernels.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* fat: Added fat parser test

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* header: fixed issue where arm64e was being returned as arm64_unknown

Co-authored-by: h3 <h3@h3s-iMac-Pro.local>

* fat: move swap_ functions to libhelper-macho.h

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* header: fix issue with CPU_SUBTYPE_ARM64_ALL mask being identified as arm64_unknown

* load-commands: added numerous load command structures

* Load Commands: Added linkedit_data, dyldinfo and entry point commands

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>

* header: print warning if header type isn't supported

* version: draft script for incrementing the build at compile time

* logger: prepend colour macros with ANSI_COLOR

Co-authored-by: h3 <h3@h3s-iMac-Pro.localdomain>
Co-authored-by: h3 <h3@h3s-iMac-Pro.local>
Co-authored-by: me@h3adsh0tzz.com <h3@h3adsh0tzzs-MacBook-Pro.local>
Co-authored-by: me@h3adsh0tzz.com <h3@h3adsh0tzzs-MBP.localdomain>