Load Command: Fix searching for load command names.
Libhelper uses an array of Load Command's and their names so a client
can fetch the String representation of the Load Command. However, the
array did not end with a NULL entry, so if the client tries to find
the last load command in the array, it won't find it.

This is now resolved.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>